requirements.txt
crewai==0.28.0
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0 